Allow processing a dynamically calculated set of class folders.
The ANT task currently can invoke FindBugs on the contents of multiple
class folders only if the folders are explicitly given using multiple
<class location="..."/> elements. If the names of the class folders are
not statically known, it is currently not possible to let ANT collect a
set of directories and pass this to the FindBugs task. To support this
use case, a nested <dirset> element is added to the task that behaves
like a nested <fileset> element for explicitly referencing multiple
class files.